transactions-api/pyproject.toml
[tool.poetry]
name = "transactions-api"
version = "0.1.0"
description = ""
authors = ["Aryan Curiel <aryan.curiel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
pydantic = "^1.8.2"
dependency-injector = "^4.34.2"
gunicorn = "^20.1.0"
pymongo = "^3.12.0"
PyJWT = "^2.1.0"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
